refactor: 댓글 및 게시물 관련 API 경로 수정
