+++ src/entities/comment/api/deleteComment.ts
@@ -1,4 +1,4 @@
-import useCommentStore from "../model/useCommentStore"
+import useCommentStore from "../../../features/comment/model/useCommentStore"
 
 // 댓글 삭제
 export const deleteComment = async (id: number, postId: number) => {
@@ -6,12 +6,12 @@ export const deleteComment = async (id: number, postId: number) => {
   try {
     await fetch(`/api/comments/${id}`, {
       method: "DELETE",
-      })
-      setComments(
-        postId,
-        comments[postId].filter((comment) => comment.id !== id),
-      )
-    } catch (error) {
+    })
+    setComments(
+      postId,
+      comments[postId].filter((comment) => comment.id !== id),
+    )
+  } catch (error) {
     console.error("댓글 삭제 오류:", error)
   }
 }
