+++ src/entities/post/api/fetchPost.ts
@@ -1,5 +1,5 @@
 import Post from "../model/type.ts"
-import User from "../../user/model/type.ts"
+import { User } from "../../user/model/type.ts"
 export interface PostParams {
   limit: number
   skip: number
@@ -19,11 +19,11 @@ export interface TagParams extends PostParams {
  */
 export const fetchPosts = async ({ limit, skip, sortBy, sortOrder }: PostParams) => {
   let url = `/api/posts?limit=${limit}&skip=${skip}`
-  
-  if (sortBy && sortBy !== 'none') {
-    url += `&sortBy=${sortBy}&order=${sortOrder || 'asc'}`
+
+  if (sortBy && sortBy !== "none") {
+    url += `&sortBy=${sortBy}&order=${sortOrder || "asc"}`
   }
-  
+
   const postsResponse = await fetch(url)
   const postsData = await postsResponse.json()
 
@@ -55,14 +55,14 @@ export const fetchPostsByTag = async ({ tag, limit, skip, sortBy, sortOrder }: T
 
   let url = `/api/posts/tag/${tag}`
   const queryParams = new URLSearchParams()
-  
-  if (limit) queryParams.append('limit', limit.toString())
-  if (skip) queryParams.append('skip', skip.toString())
-  if (sortBy && sortBy !== 'none') {
-    queryParams.append('sortBy', sortBy)
-    queryParams.append('order', sortOrder || 'asc')
+
+  if (limit) queryParams.append("limit", limit.toString())
+  if (skip) queryParams.append("skip", skip.toString())
+  if (sortBy && sortBy !== "none") {
+    queryParams.append("sortBy", sortBy)
+    queryParams.append("order", sortOrder || "asc")
   }
-  
+
   if (queryParams.toString()) {
     url += `?${queryParams.toString()}`
   }
@@ -91,13 +91,13 @@ export const fetchPostsByTag = async ({ tag, limit, skip, sortBy, sortOrder }: T
  * @param query 검색 쿼리
  * @returns 게시물 목록과 총 게시물 수
  */
-export const searchPosts = async (query: string, options?: { sortBy?: string, sortOrder?: string }) => {
+export const searchPosts = async (query: string, options?: { sortBy?: string; sortOrder?: string }) => {
   let url = `/api/posts/search?q=${query}`
-  
-  if (options?.sortBy && options.sortBy !== 'none') {
-    url += `&sortBy=${options.sortBy}&order=${options.sortOrder || 'asc'}`
+
+  if (options?.sortBy && options.sortBy !== "none") {
+    url += `&sortBy=${options.sortBy}&order=${options.sortOrder || "asc"}`
   }
-  
+
   const response = await fetch(url)
   const data = await response.json()
   return data
