+++ src/features/comment/api/index.ts
@@ -1,0 +1,3 @@
+export * from "../../../entities/comment/api/updateComment"
+export * from "../../../entities/comment/api/deleteComment"
+export * from "../../../entities/comment/api/likeComment"
