--- src/features/comments/api/index.ts
@@ -1,3 +1,0 @@
-export * from "./updateComment"
-export * from "./deleteComment"
-export * from "./likeComment"
