+++ src/features/posts/api/index.ts
@@ -1,2 +1,2 @@
-export * from "./useDeletePost"
-export * from "./useUpdatePost"
+export * from "../../../entities/post/api/useDeletePost"
+export * from "../../../entities/post/api/useUpdatePost"
