+++ src/pages/PostsManagerPage.tsx
@@ -5,15 +5,15 @@ import EditPostDialog from "../features/posts/ui/EditPostDialog.tsx"
 import UserModal from "../widgets/userModal/UserModal.tsx"
 import PostsPagination from "../features/filters/ui/PostPagination.tsx"
 import FilterWrapper from "../features/filters/ui/filters/FilterWrapper.tsx"
-import AddCommentDialog from "../features/comments/ui/AddCommentDialog.tsx"
-import EditCommentDialog from "../features/comments/ui/EditCommentDialog.tsx"
+import AddCommentDialog from "../features/comment/ui/AddCommentDialog.tsx"
+import EditCommentDialog from "../features/comment/ui/EditCommentDialog.tsx"
 import PostsTable from "../features/posts/ui/table/PostsTable.tsx"
 // Store
 import usePostsStore from "../features/posts/model/usePostsStore.ts"
 import useGlobalStore from "../shared/model/useGlobalStore.ts"
 import useFilterStore from "../features/filters/model/useFilterStore.ts"
 import { useURLParams } from "../shared/lib/hooks/useURLParams.ts"
-import { usePostsQuery, useTagsQuery } from "../features/posts/model/queries.ts"
+import { usePostsQuery, useTagsQuery } from "../entities/post/model/queries.ts"
 import PostDetailDialog from "../widgets/PostDetailDialog/PostDetailDialog.tsx"
 import PostsHeader from "../features/posts/ui/PostHeader.tsx"
 // UI
