+++ src/widgets/PostDetailDialog/PostDetailDialog.tsx
@@ -1,12 +1,12 @@
 import { Dialog, DialogContent, DialogHeader, DialogTitle } from "../../shared/ui/Dialog"
 import usePostsStore from "../../features/posts/model/usePostsStore"
 import useFilterStore from "../../features/filters/model/useFilterStore"
-import useCommentStore from "../../features/comments/model/useCommentStore"
-import CommentsList from "../../features/comments/ui/CommentsList"
+import useCommentStore from "../../features/comment/model/useCommentStore"
+import CommentsList from "../../features/comment/ui/CommentsList"
 import highlightText from "../../shared/lib/util/highlightText"
 import { fetchComments } from "../../entities/comment/api/fetchComments"
-import { likeComment } from "../../features/comments/api/likeComment"
-import { deleteComment } from "../../features/comments/api/deleteComment"
+import { likeComment } from "../../entities/comment/api/likeComment"
+import { deleteComment } from "../../entities/comment/api/deleteComment"
 import { useShallow } from "zustand/shallow"
 import { useEffect } from "react"
